+++ src/io/psse.jl
@@ -127,6 +127,8 @@ Parses PSS(R)E-style Branch data into a PowerModels-style Dict. "source_id" is
 given by `["I", "J", "CKT"]` in PSS(R)E Branch specification.
 """
 function psse2pm_branch!(pm_data::Dict, pti_data::Dict, import_all::Bool)
+    
+
     pm_data["branch"] = []
     if haskey(pti_data, "BRANCH")
         for (i, branch) in enumerate(pti_data["BRANCH"])
@@ -154,6 +156,16 @@ function psse2pm_branch!(pm_data::Dict, pti_data::Dict, import_all::Bool)
             sub_data["index"] = i
 
             import_remaining!(sub_data, branch, import_all; exclude=["B", "BI", "BJ"])
+
+            if sub_data["rate_a"] == 0.0
+                delete!(sub_data, "rate_a")
+            end
+            if sub_data["rate_b"] == 0.0
+                delete!(sub_data, "rate_b")
+            end
+            if sub_data["rate_c"] == 0.0
+                delete!(sub_data, "rate_c")
+            end
 
             push!(pm_data["branch"], sub_data)
         end
@@ -380,6 +392,16 @@ function psse2pm_transformer!(pm_data::Dict, pti_data::Dict, import_all::Bool)
                 sub_data["rate_b"] = pop!(transformer, "RATB1")
                 sub_data["rate_c"] = pop!(transformer, "RATC1")
 
+                if sub_data["rate_a"] == 0.0
+                    delete!(sub_data, "rate_a")
+                end
+                if sub_data["rate_b"] == 0.0
+                    delete!(sub_data, "rate_b")
+                end
+                if sub_data["rate_c"] == 0.0
+                    delete!(sub_data, "rate_c")
+                end
+
                 sub_data["tap"] = pop!(transformer, "WINDV1") / pop!(transformer, "WINDV2")
                 sub_data["shift"] = pop!(transformer, "ANG1")
 
@@ -464,6 +486,16 @@ function psse2pm_transformer!(pm_data::Dict, pti_data::Dict, import_all::Bool)
                     sub_data["rate_a"] = pop!(transformer, "RATA$m")
                     sub_data["rate_b"] = pop!(transformer, "RATB$m")
                     sub_data["rate_c"] = pop!(transformer, "RATC$m")
+
+                    if sub_data["rate_a"] == 0.0
+                        delete!(sub_data, "rate_a")
+                    end
+                    if sub_data["rate_b"] == 0.0
+                        delete!(sub_data, "rate_b")
+                    end
+                    if sub_data["rate_c"] == 0.0
+                        delete!(sub_data, "rate_c")
+                    end
 
                     sub_data["tap"] = pop!(transformer, "WINDV$m")
                     sub_data["shift"] = pop!(transformer, "ANG$m")
@@ -630,7 +662,6 @@ function psse2pm_dcline!(pm_data::Dict, pti_data::Dict, import_all::Bool)
 end
 
 
-
 """
     parse_psse(pti_data)
 
@@ -659,10 +690,11 @@ function parse_psse(pti_data::Dict; import_all=false)::Dict
     psse2pm_transformer!(pm_data, pti_data, import_all)
     psse2pm_dcline!(pm_data, pti_data, import_all)
 
-    import_remaining!(pm_data, pti_data, import_all; exclude=["CASE IDENTIFICATION", "BUS", "LOAD",
-                                                              "FIXED SHUNT", "SWITCHED SHUNT", "GENERATOR",
-                                                              "BRANCH", "TRANSFORMER", "TWO-TERMINAL DC",
-                                                              "VOLTAGE SOURCE CONVERTER"])
+    import_remaining!(pm_data, pti_data, import_all; exclude=[
+        "CASE IDENTIFICATION", "BUS", "LOAD", "FIXED SHUNT",
+        "SWITCHED SHUNT", "GENERATOR","BRANCH", "TRANSFORMER",
+        "TWO-TERMINAL DC", "VOLTAGE SOURCE CONVERTER"
+    ])
 
     # update lookup structure
     for (k, v) in pm_data